// client/@core/route-manifest.ts
// Auto-generated by generateManifest
// DO NOT EDIT MANUALLY

export const routeTree = [
  {
    "path": "/",
    "children": [
      {
        "path": "/about",
        "children": [
          {
            "path": "/about/:id",
            "pagePath": "../@pages/about/[id]/page"
          }
        ],
        "pagePath": "../@pages/about/page"
      },
      {
        "path": "/landingPage",
        "pagePath": "../@pages/landingPage/page"
      }
    ],
    "layoutPath": "../@pages/layout",
    "pagePath": "../@pages/page"
  },
  {
    "path": "/admin",
    "layoutPath": "../@admin/layout",
    "pagePath": "../@admin/page"
  },
  {
    "path": "*",
    "pagePath": "../404.tsx"
  }
];

export const moduleMap = {
  "/": () => import("../@pages/page"),
  "/@layout": () => import("../@pages/layout"),
  "/about": () => import("../@pages/about/page"),
  "/about/:id": () => import("../@pages/about/[id]/page"),
  "/landingPage": () => import("../@pages/landingPage/page"),
  "/admin": () => import("../@admin/page"),
  "/admin@layout": () => import("../@admin/layout"),
  "*": () => import("../404.tsx"),
};